Refactor query syntax and extract pagination to Outlet component
 - Update query syntax from `-tag:"hide_from_combine"` to `tag_not:"hide-from-combine"` in artist treatment loader.
 - Remove PaginationContent and related code from treatments route, moving pagination handling to an Outlet component.
 - Remove unused imports and fragments related to pagination and user collection.
